+++ crypto/encoding/codec.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/cometbft/cometbft/crypto"
+	"github.com/cometbft/cometbft/crypto/bls12381"
 	"github.com/cometbft/cometbft/crypto/ed25519"
 	"github.com/cometbft/cometbft/crypto/secp256k1"
 	"github.com/cometbft/cometbft/libs/json"
@@ -35,6 +36,9 @@ func init() {
 	json.RegisterType((*pc.PublicKey)(nil), "tendermint.crypto.PublicKey")
 	json.RegisterType((*pc.PublicKey_Ed25519)(nil), "tendermint.crypto.PublicKey_Ed25519")
 	json.RegisterType((*pc.PublicKey_Secp256K1)(nil), "tendermint.crypto.PublicKey_Secp256K1")
+	if bls12381.Enabled {
+		json.RegisterType((*pc.PublicKey_Bls12381)(nil), "tendermint.crypto.PublicKey_Bls12381")
+	}
 }
 
 // PubKeyToProto takes crypto.PubKey and transforms it to a protobuf Pubkey
@@ -51,6 +55,16 @@ func PubKeyToProto(k crypto.PubKey) (pc.PublicKey, error) {
 		kp = pc.PublicKey{
 			Sum: &pc.PublicKey_Secp256K1{
 				Secp256K1: k,
+			},
+		}
+	case bls12381.PubKey:
+		if !bls12381.Enabled {
+			return kp, ErrUnsupportedKey{Key: k}
+		}
+
+		kp = pc.PublicKey{
+			Sum: &pc.PublicKey_Bls12381{
+				Bls12381: k,
 			},
 		}
 	default:
@@ -77,6 +91,21 @@ func PubKeyFromProto(k pc.PublicKey) (crypto.PubKey, error) {
 		}
 		pk := make(secp256k1.PubKey, secp256k1.PubKeySize)
 		copy(pk, k.Secp256K1)
+		return pk, nil
+	case *pc.PublicKey_Bls12381:
+		if !bls12381.Enabled {
+			return nil, ErrUnsupportedKey{Key: k}
+		}
+
+		if len(k.Bls12381) != bls12381.PubKeySize {
+			return nil, ErrInvalidKeyLen{
+				Key:  k,
+				Got:  len(k.Bls12381),
+				Want: bls12381.PubKeySize,
+			}
+		}
+		pk := make(bls12381.PubKey, bls12381.PubKeySize)
+		copy(pk, k.Bls12381)
 		return pk, nil
 	default:
 		return nil, fmt.Errorf("fromproto: key type %v is not supported", k)