+++ types/block.go
@@ -590,11 +590,12 @@ const (
 )
 
 const (
-	// Max size of commit without any commitSigs -> 82 for BlockID, 8 for Height, 4 for Round.
+	// MaxCommitOverheadBytes is the max size of commit without any commitSigs -> 82 for BlockID, 8 for Height, 4 for Round.
 	MaxCommitOverheadBytes int64 = 94
-	// Commit sig size is made up of 64 bytes for the signature, 20 bytes for the address,
-	// 1 byte for the flag and 14 bytes for the timestamp
-	MaxCommitSigBytes int64 = 109
+
+	// MaxCommitSigBytes is the commit sig size is made up of 96 bytes for the signature, 20 bytes for the address,
+	// 1 byte for the flag and 14 bytes for the timestamp.
+	MaxCommitSigBytes int64 = 131 + 10 // where's the 10 from?
 )
 
 // CommitSig is a part of the Vote included in a Commit.
@@ -606,8 +607,9 @@ type CommitSig struct {
 }
 
 func MaxCommitBytes(valCount int) int64 {
+	// protoEncodingOverhead represents the overhead in bytes when encoding a protocol buffer message.
+	const protoEncodingOverhead int64 = 3
 	// From the repeated commit sig field
-	var protoEncodingOverhead int64 = 2
 	return MaxCommitOverheadBytes + ((MaxCommitSigBytes + protoEncodingOverhead) * int64(valCount))
 }